src/requirements.txt
sanic~=22.6.2
python-dotenv~=0.20.0
SQLAlchemy~=1.4.40
alembic~=1.8.1
pydantic~=1.9.2
sanic-jwt-extended~=1.0.dev12